refactor(graph): extract store contracts
